+++ src/utils/FileRecordTools/Records/BedPlusInterval.h
@@ -29,7 +29,7 @@ public:
 	virtual FileRecordTypeChecker::RECORD_TYPE getType() const { return FileRecordTypeChecker::BED_PLUS_RECORD_TYPE; }
 
 	virtual const string &getField(int fieldNum) const;
-	virtual int getNumFields() const  { return _numFixedFields + _plusFields.size(); }
+	virtual int getNumFields() const  { return (int)_numFixedFields + (int)_plusFields.size(); }
 
 	virtual void setNumPrintFields(int num) { _numPrintFields = num; }
 	virtual int getNumPrintFields() const { return _numPrintFields; }
